Fixed the test files

+++ estore-api/src/main/java/com/estore/api/estoreapi/model/Ingredient.java
@@ -22,9 +22,9 @@ public class Ingredient {
     @JsonProperty("type")
     private String type;
     @JsonProperty("price")
-    private float price;
+    private double price;
     @JsonProperty("volume")
-    private float volume;
+    private double volume;
 
     /**
      * Create a ingredient with the given id, name, type, price, and volume
@@ -44,7 +44,7 @@ public class Ingredient {
      *               value, i.e. 0 for int
      */
     public Ingredient(@JsonProperty("id") int id, @JsonProperty("name") String name, @JsonProperty("type") String type,
-            @JsonProperty("price") float price, @JsonProperty("volume") float volume) {
+            @JsonProperty("price") double price, @JsonProperty("volume") double volume) {
         this.id = id;
         this.name = name;
         this.type = type;
@@ -105,7 +105,7 @@ public class Ingredient {
      * 
      * @param price The price of the ingredient
      */
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 
@@ -114,7 +114,7 @@ public class Ingredient {
      * 
      * @return The price of the ingredient
      */
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
@@ -124,7 +124,7 @@ public class Ingredient {
      * 
      * @param volume The volume of the ingredient
      */
-    public void setVolume(float volume) {
+    public void setVolume(double volume) {
         this.volume = volume;
     }
 
@@ -133,7 +133,7 @@ public class Ingredient {
      * 
      * @return The volume of the ingredient
      */
-    public float getVolume() {
+    public double getVolume() {
         return volume;
     }
